Include unit tests for pagination/records.rs resources inside the module
